Added breakout logic
In the event that this is run from a thread, the use of the
continueAcceptLoop method in the while loop, in conjunction with
serverSock's setSoTimeout method, allows the loop to cancel or break out
to complete listen() should the prerequisites be met.

As for those prerequisites, those are determined by the extending class.
continueAcceptLoop is abstract.